update pending bookings page
